refactor(dashboard-v2): move upgrade screen under /payments url
